+++ netports/__init__.py
@@ -3,13 +3,14 @@
 from netports.exceptions import NetportsValueError
 from netports.intf import Intf
 from netports.intf_gm import IntfGM, intfrange
+from netports.intf_map import short_to_long, short_to_short, long_to_short, long_to_long
 from netports.ip import IP_NAMES, IP_NUMBERS, iip, sip, ip_pairs
 from netports.item import Item
 from netports.ports import inumbers, parse_range, snumbers
 from netports.range import Range
 from netports.tcp import stcp, itcp
 from netports.vlan import ivlan, svlan
-from netports.intf_map import short_to_long, short_to_short, long_to_short, long_to_long
+
 __all__ = [
     "IP_NAMES",
     "IP_NUMBERS",