+++ netports/intf_map.py
@@ -1,5 +1,4 @@
 """Interface name mapping"""
-from netports.static import PLATFORMS
 from netports.types_ import DStr
 
 MAP_OTHER = {  # parent of ios
@@ -18,6 +17,14 @@ MAP_OTHER = {  # parent of ios
     "Vi": "Virtual-Access",
     "Vt": "Virtual-Template",
 }
+MAP_CISCO_ASR = {
+    "Te": "TenGigE",  # overlapped: ios
+    "Hu": "HundredGigE",
+    "BE": "Bundle-Ether",
+    "Tu": "tunnel-ip",  # low-priority, overlapped: ios, nxos
+    "ti": "tunnel-ip",
+    "Mg": "MgmtEth",
+}
 MAP_CISCO_IOS = {  # parent of nxos, asr
     "Eth": "Ethernet",
     "Fa": "FastEthernet",
@@ -33,14 +40,6 @@ MAP_CISCO_NXOS = {
     "Lo": "loopback",  # overlapped: cisco, asr
     "mgmt": "mgmt",
 }
-MAP_CISCO_ASR = {
-    "Te": "TenGigE",  # overlapped: ios
-    "Hu": "HundredGigE",
-    "BE": "Bundle-Ether",
-    "Tu": "tunnel-ip",  # low-priority, overlapped: ios, nxos
-    "ti": "tunnel-ip",
-    "Mg": "MgmtEth",
-}
 MAP_HP_COMWARE = {  # h3c
     "GE": "GigabitEthernet",
     "Te": "Ten-GigabitEthernet",  # low-priority, overlapped: cisco
@@ -51,53 +50,65 @@ MAP_HP_COMWARE = {  # h3c
 MAP_HP_PROCURVE = {  # hpc
     "Trk": "Trk",
 }
+ALL_SHORT = sorted({
+    *MAP_OTHER,
+    *MAP_CISCO_IOS,
+    *MAP_CISCO_NXOS,
+    *MAP_CISCO_ASR,
+    *MAP_HP_COMWARE,
+    *MAP_HP_PROCURVE,
+})
 
 
-def long_to_short(platform: str = "", key_lower: bool = False, value_lower: bool = False) -> DStr:
-    """Returns Interfaces map long-to-short, platform specific
+def long_to_short(device_type: str = "",
+                  key_lower: bool = False,
+                  value_lower: bool = False) -> DStr:
+    """Returns Interfaces map long-to-short, device_type specific
     ::
-        :param platform: Platform, increase priority of platform specific keys
+        :param device_type: Netmiko device type, increase priority of device_type specific keys
         :param key_lower: True - keys lower-case, False - keys upper-case
         :param value_lower: True - values lower-case, False - values upper-case
         :return: Interfaces map
         :example:
             long_to_short() -> {"Vlan": "V", ...}
-            long_to_short(platform="cisco_ios") -> {"Vlan": "Vlan", ...}
-            long_to_short(platform="cisco_ios", key_lower=True) -> {"vlan": "Vlan", ...}
+            long_to_short(device_type="cisco_ios") -> {"Vlan": "Vlan", ...}
+            long_to_short(device_type="cisco_ios", key_lower=True) -> {"vlan": "Vlan", ...}
     """
-    data: DStr = short_to_long(platform=platform)
+    data: DStr = short_to_long(device_type=device_type)
     data = {v: k for k, v in data.items()}
     data_ = _overlapped(data)
     data.update(data_)
 
-    if platform == "cisco_asr":
+    if device_type == "cisco_asr":
         data["Tunnel"] = "ti"
-    if platform == "hp_comware":
+    if device_type == "hp_comware":
         data["TenGigabitEthernet"] = "XGE"
 
     data = _lower(data, key_lower, value_lower)
     return data
 
 
-def long_to_long(platform: str = "", key_lower: bool = False, value_lower: bool = False) -> DStr:
-    """Returns Interfaces map long-to-long, platform specific
+def long_to_long(device_type: str = "", key_lower: bool = False, value_lower: bool = False) -> DStr:
+    """Returns Interfaces map long-to-long, device_type specific
     ::
-        :param platform: Platform, increase priority of platform specific keys
+        :param device_type: Netmiko device type, increase priority of device_type specific keys
         :param key_lower: True - keys lower-case, False - keys upper-case
         :param value_lower: True - values lower-case, False - values upper-case
         :return: Interfaces map
     """
-    data: DStr = short_to_long(platform=platform)
+    data: DStr = short_to_long(device_type=device_type)
     data = {v: k for k, v in data.items()}
     data = {k: k for k in data}
     data = _lower(data, key_lower, value_lower)
     return data
 
 
-def short_to_long(platform: str = "", key_lower: bool = False, value_lower: bool = False) -> DStr:
-    """Returns Interfaces map short-to-long, platform specific
+def short_to_long(device_type: str = "",
+                  key_lower: bool = False,
+                  value_lower: bool = False) -> DStr:
+    """Returns Interfaces map short-to-long, device_type specific
     ::
-        :param platform: Platform, increase priority of platform specific keys
+        :param device_type: Netmiko device type, increase priority of device_type specific keys
         :param key_lower: True - keys lower-case, False - keys upper-case
         :param value_lower: True - values lower-case, False - values upper-case
         :return: Interfaces map
@@ -106,27 +117,27 @@ def short_to_long(platform: str = "", key_lower: bool = False, value_lower: bool
             short_to_long(key_lower=True) -> {"fa": "FastEthernet", ...}
             short_to_long(value_lower=True) -> {"Fa": "fastethernet", ...}
     """
-    if platform == "cisco_asr":
+    if device_type == "cisco_asr":
         data = MAP_OTHER.copy()
         data.update(MAP_CISCO_IOS)
         data.update(MAP_CISCO_ASR)
-    elif platform == "cisco_ios":
+    elif device_type == "cisco_ios":
         data = MAP_OTHER.copy()
         data.update(MAP_CISCO_IOS)
-    elif platform == "cisco_nxos":
+    elif device_type == "cisco_nxos":
         data = MAP_OTHER.copy()
         data.update(MAP_CISCO_IOS)
         data.update(MAP_CISCO_NXOS)
-    elif platform == "hp_comware":  # h3c
+    elif device_type == "hp_comware":  # h3c
         data = MAP_OTHER.copy()
         data.update(MAP_CISCO_IOS)
         data.update(MAP_HP_COMWARE)
-    elif platform == "hp_procurve":  # h3c
+    elif device_type == "hp_procurve":  # h3c
         data = MAP_OTHER.copy()
         data.update(MAP_CISCO_IOS)
         data.update(MAP_HP_PROCURVE)
     else:
-        data: DStr = MAP_HP_COMWARE.copy()
+        data = MAP_HP_COMWARE.copy()
         data.update(MAP_HP_PROCURVE)
         data.update(MAP_CISCO_ASR)
         data.update(MAP_CISCO_NXOS)
@@ -136,10 +147,12 @@ def short_to_long(platform: str = "", key_lower: bool = False, value_lower: bool
     return data
 
 
-def short_to_short(platform: str = "", key_lower: bool = False, value_lower: bool = False) -> DStr:
-    """Returns Interfaces map short-to-short, platform specific
+def short_to_short(device_type: str = "",
+                   key_lower: bool = False,
+                   value_lower: bool = False) -> DStr:
+    """Returns Interfaces map short-to-short, device_type specific
     ::
-        :param platform: Platform, increase priority of platform specific keys
+        :param device_type: Netmiko device type, increase priority of device_type specific keys
         :param key_lower: True - keys lower-case, False - keys upper-case
         :param value_lower: True - values lower-case, False - values upper-case
         :return: Interfaces map
@@ -147,10 +160,10 @@ def short_to_short(platform: str = "", key_lower: bool = False, value_lower: boo
             short_to_short(key_lower=True) -> {"fa": "Fa", ...}
             short_to_short(value_lower=True) -> {"Fa": "fa", ...}
     """
-    data: DStr = short_to_long(platform=platform)
+    data: DStr = short_to_long(device_type=device_type)
     data = {k: k for k in data}
-    intf_map_l2s: DStr = long_to_short(platform=platform)
-    intf_map_s2l: DStr = short_to_long(platform=platform)
+    intf_map_l2s: DStr = long_to_short(device_type=device_type)
+    intf_map_s2l: DStr = short_to_long(device_type=device_type)
     for short_upper, long_upper in intf_map_s2l.items():
         if value := intf_map_l2s.get(long_upper):
             data[short_upper] = value
@@ -158,12 +171,10 @@ def short_to_short(platform: str = "", key_lower: bool = False, value_lower: boo
     return data
 
 
-
-
 # ============================== helper ==============================
 
 def _overlapped(data: DStr) -> DStr:
-    """Returns Interfaces map long-to-short that overlapped with short key in other platforms
+    """Returns Interfaces map long-to-short that overlapped with short key in other device_type
 
     """
     data_ = {v: k for k, v in MAP_OTHER.items()}