+++ netports/static.py
@@ -1,6 +1,12 @@
 """Static settings"""
 
-PLATFORMS = ["cisco_asr", "cisco_ios", "cisco_nxos", "hp_comware", "hp_procurve"]
+DEVICE_TYPES = [
+    "cisco_asr",
+    "cisco_ios",
+    "cisco_nxos",
+    "hp_comware",
+    "hp_procurve",
+]
 RANGE_SPLITTER = "-"
 RANGE_SPLITTER_HPE = " to "
 RANGE_SPLITTERS = [RANGE_SPLITTER, RANGE_SPLITTER_HPE]