+++ netports/types_.py
@@ -1,6 +1,7 @@
 """Typing"""
 from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Union
 
+# one-level
 DAny = Dict[str, Any]
 DStr = Dict[str, str]
 DiAny = Dict[int, Any]
@@ -13,14 +14,16 @@ OInt = Optional[int]
 SInt = Set[int]
 SStr = Set[str]
 StrInt = Union[str, int]
-TIntStr = Tuple[int, str]
-
-DLStr = Dict[str, LStr]
-DSStr = Dict[str, SStr]
-IStrInt = Union[IStr, IInt]
-LTIntStr = List[TIntStr]
-StrIInt = Union[str, int, IInt]
-T2SInt = Tuple[SInt, SInt]
 T2Str = Tuple[str, str]
 T3Str = Tuple[str, str, str]
 T5Str = Tuple[str, str, str, str, str]
+TIntStr = Tuple[int, str]
+
+# two-level
+DLStr = Dict[str, LStr]
+DSStr = Dict[str, SStr]
+IStrInt = Union[IStr, IInt]
+LT2Str = List[T2Str]
+LTIntStr = List[TIntStr]
+StrIInt = Union[str, int, IInt]
+T2SInt = Tuple[SInt, SInt]